+++ src/lib.rs
@@ -1,12 +1,10 @@
 use crate::fmt_md::MdOptions;
 use crate::fmt_md_inlines::MdInlinesWriterOptions;
 use crate::fmt_plain::PlainOutputOpts;
+use crate::md_elem::{InvalidMd, MdDoc, MdElemRef, ReadOptions, SerdeDoc};
 use crate::output::{OutputOpts, Stream};
 use crate::query::ParseError;
 use crate::select::{SelectorAdapter, SelectorChain};
-use crate::tree::{InvalidMd, MdDoc, ReadOptions};
-use crate::tree_ref::MdElemRef;
-use crate::tree_ref_serde::SerdeDoc;
 use cli::{Cli, OutputFormat};
 use output::Output;
 use pest::error::ErrorVariant;
@@ -26,14 +24,11 @@ mod fmt_str;
 mod footnote_transform;
 mod link_transform;
 mod matcher;
+pub mod md_elem;
 mod output;
 mod query;
 mod select;
 mod str_utils;
-mod tree;
-mod tree_ref;
-mod tree_ref_serde;
-mod tree_test_utils;
 mod utils_for_test;
 mod vec_utils;
 mod words_buffer;
