+++ src/link_transform.rs
@@ -1,6 +1,6 @@
 use crate::fmt_md_inlines::{LinkLike, MdInlinesWriter, MdInlinesWriterOptions};
+use crate::md_elem::*;
 use crate::output::Output;
-use crate::tree::{Inline, LinkReference, MdContext};
 use clap::ValueEnum;
 use std::borrow::Cow;
 use std::collections::hash_map::Entry;
@@ -204,8 +204,6 @@ fn inlines_to_string<'md>(inline_writer: &mut MdInlinesWriter<'md>, inlines: &'m
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tree::{Link, LinkDefinition, MdContext, Text, TextVariant};
-    use crate::tree_test_utils::*;
     use crate::utils_for_test::*;
 
     enum Combo {
@@ -300,7 +298,6 @@ mod tests {
 
     mod reference {
         use super::*;
-        use crate::tree::{Formatting, FormattingVariant};
 
         #[test]
         fn inline() {
