+++ src/select/sel_link_like.rs
@@ -1,7 +1,7 @@
 use crate::matcher::StringMatcher;
+use crate::md_elem::*;
 use crate::select::match_selector::MatchSelector;
 use crate::select::LinklikeMatcher;
-use crate::tree::{Image, Link};
 
 #[derive(Debug, PartialEq)]
 pub struct LinkSelector {
